- updated the makefiles and code to allow compliation with MinGW once again (some of patch #875485 by perditionc) - fixed errors and warnings given by gcc
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@3513 212acab6-be3b-0410-9dea-997c60f758d6

+++ Source/DialogTemplate.cpp
@@ -64,7 +64,7 @@ void ReadVarLenArr(BYTE* &seeker, char* &readInto, unsigned int uCodePage) {
     if (IS_INTRESOURCE(x)) { \
       *(WORD*)seeker = 0xFFFF; \
       seeker += sizeof(WORD); \
-      *(WORD*)seeker = WORD(x); \
+      *(WORD*)seeker = WORD(DWORD(x)); \
       seeker += sizeof(WORD); \
     } \
     else { \
